Tehtävät 4.15-4.21

+++ blogilista-backend/controllers/blogs.js
@@ -1,10 +1,15 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog
+    .find({}).populate('user', {username: 1, name: 1})
+
   response.json(blogs.map(blog => blog.toJSON()))
 })
+
 
 blogsRouter.post('/', async (request, response, next) => {
   const body = request.body
@@ -13,15 +18,25 @@ blogsRouter.post('/', async (request, response, next) => {
     return response.status(400).json({ error: 'title or url missing' })
   }
 
-  const blog = new Blog({
-    title: body.title,
-    author: body.author,
-    url: body.url,
-    likes: body.likes === undefined ? 0 : body.likes
-  })
+  try {
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!request.token || !decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+    
+    const user = await User.findById(body.userId)
 
-  try { 
+    const blog = new Blog({
+      title: body.title,
+      author: body.author,
+      url: body.url,
+      likes: body.likes === undefined ? 0 : body.likes,
+      user: user._id
+    })
+    
     const savedBlog = await blog.save()
+    user.blogs = user.blogs.concat(savedBlog._id)
+    await user.save()
     response.json(savedBlog.toJSON())
   } catch(exception) {
     next(exception)
@@ -29,6 +44,13 @@ blogsRouter.post('/', async (request, response, next) => {
 })
 
 blogsRouter.delete('/:id', async (request, response, next) => {
+  const blog = await Blog.findById(request.params.id)
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  
+  if (blog.user.toString() !== decodedToken.id) {
+    return response.status(401).json({ error: 'only the user that added the blog can remove it'})
+  }
+  
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
 })
